+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* feat: add static code diagnostics `avoid-unnecessary-type-casts`
 * fix: fixed issue with type check in `avoid-unnecessary-type-assertions`
 * feat: introduce file metrics
 * feat: add static code diagnostics `avoid-unnecessary-type-assertions`